refactor: rename lib/remark to lib/markdown
