+++ next.config.mjs
@@ -8,8 +8,8 @@ import { remarkCodeHike } from "@code-hike/mdx";
 import { readFile } from "fs/promises";
 import rehypeAutolinkHeadings from "rehype-autolink-headings";
 import { h } from "hastscript";
-import { remarkInfoBlock } from "./src/lib/remark/remark-info-block.mjs";
-import { remarkYoutube } from "./src/lib/remark/remark-youtube.mjs";
+import { remarkInfoBlock } from "./src/lib/markdown/remark-info-block.mjs";
+import { remarkYoutube } from "./src/lib/markdown/remark-youtube.mjs";
 import {
   infoBlockHeader,
   infoBlockChildren,
