+++ src/pages/docs/start-here/getting-started.mdx
@@ -96,7 +96,9 @@ Jump right in
   - [pipeline-backend](https://github.com/instill-ai/pipeline-backend), [connector-backend](https://github.com/instill-ai/connector-backend), [model-backend](https://github.com/instill-ai/model-backend), [mgmt-backend](https://github.com/instill-ai/mgmt-backend), [console](https://github.com/instill-ai/console), etc., for specific issues.
 - Please refer to the [VDP project board](https://github.com/orgs/instill-ai/projects/5) to track progress.
 
-> Code in the main branch tracks under-development progress towards the next release and may not work as expected. If you are looking for a stable alpha version, please use [latest release](https://github.com/instill-ai/vdp/releases).
+:::info{type=warning}
+Code in the main branch tracks under-development progress towards the next release and may not work as expected. If you are looking for a stable alpha version, please use [latest release](https://github.com/instill-ai/vdp/releases).
+:::
 
 ## Community support
